Refactor Event. Extract common code into AbstractEvent class
